northd: fix infinite loop in ovn_allocate_tnlid()
In case if all tunnel ids are exhausted, ovn_allocate_tnlid() function
iterates over tnlids indefinitely when *hint < min.  This is because
when tnlid reaches max, next tnlid is min and for-loop never reaches exit
condition for tnlid != *hint.

This patch sets a starting hint value to OVN_MIN_DP_KEY_LOCAL (1).

The downside of this change is that first allocated tnlid becomes 2.
This shouldn't be a problem.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>