northd: fix infinite loop in ovn_allocate_tnlid()
In case if all tunnel ids are exhausted, ovn_allocate_tnlid() function
iterates over tnlids indefinitely when *hint is outside of [min, max].
This is because when tnlid reaches max, next tnlid is min and for-loop
never reaches exit condition for tnlid != *hint.

This patch fixes mentioned issue and adds a testcase.

Signed-off-by: Vladislav Odintsov <odivlad@gmail.com>
Acked-by: Mark Michelson <mmichels@redhat.com>